--- src/pages/mypage/chat/message/edit/store/index.ts
@@ -1,11 +1,0 @@
-const messageState = {
-    id: "",
-    threadId: "",
-    userId: 0,
-    iconUrl: "",
-    message: "",
-    createdAt: new Date(),
-    updatedAt: new Date(),
-}
-
-export default messageState
