--- src/pages/mypage/chat/message/index/store/index.tsx
@@ -1,14 +1,0 @@
-
-import firebase from "firebase/app"
-
-const messageState = [{
-    id: "",
-    threadId: 0,
-    userId: 0,
-    iconUrl: "",
-    message: "",
-    createdAt: new Date(),
-    updatedAt: new Date(),
-}]
-
-export default messageState
