--- src/pages/mypage/chat/message/new/store/index.ts
@@ -1,10 +1,0 @@
-const messageState = {
-    threadId: "",
-    userId: 0,
-    iconUrl: "",
-    message: "",
-    createdAt: new Date(),
-    updatedAt: new Date(),
-}
-
-export default messageState
