+++ src/pages/mypage/chat/store/messageState.ts
@@ -1,0 +1,29 @@
+export const existMessages = [{
+    id: "",
+    threadId: "",
+    userId: 0,
+    iconUrl: "",
+    message: "",
+    createdAt: new Date(),
+    updatedAt: new Date(),
+}]
+
+export const existMessage = {
+    id: "",
+    threadId: "",
+    userId: 0,
+    iconUrl: "",
+    message: "",
+    createdAt: new Date(),
+    updatedAt: new Date(),
+}
+
+export const message = {
+    threadId: "",
+    userId: 0,
+    iconUrl: "",
+    message: "",
+    createdAt: new Date(),
+    updatedAt: new Date(),
+}
+
