+++ src/pages/mypage/chat/store/threadState.ts
@@ -1,0 +1,17 @@
+export const threads = [{
+    id: "",
+    transactionId: 0,
+    title: "",
+    description: "",
+    createdAt: new Date(),
+    updatedAt: new Date(),
+}]
+
+export const thread = {
+    id: "",
+    transactionId: 0,
+    title: "",
+    description: "",
+    createdAt: new Date(),
+    updatedAt: new Date(),
+}
