--- src/pages/mypage/chat/thread/index/store/index.ts
@@ -1,12 +1,0 @@
-import firebase from "firebase/app"
-
-const threadsState = [{
-    id: "",
-    transactionId: 0,
-    title: "",
-    description: "",
-    createdAt: new Date(),
-    updatedAt: new Date(),
-}]
-
-export default threadsState
