--- src/pages/mypage/chat/thread/new/store/index.tsx
@@ -1,11 +1,0 @@
-import firebase from "firebase/app"
-
-const threadState = {
-    transactionId: 0,
-        title: "",
-        description: "",
-        createdAt: new Date(),
-        updatedAt: new Date(),
-}
-
-export default threadState
